LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/option-list-expected.txt


FAIL select's option list should not include descendants of options, hrs, or nested optgroups. assert_equals: childSelect.length expected 3 but got 2
